backend/requirements.txt
Django>=5.2.8
psycopg2-binary>=2.9.9
python-decouple>=3.8
dj-database-url>=2.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
gunicorn>=21.2.0
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0
faker>=20.0.0
bcrypt>=4.1.0
requests>=2.31.0

# MongoDB
pymongo>=4.6.0
dnspython>=2.4.0

# AWS S3
boto3>=1.34.0
